Fix bug and add IPadTests scheme
