+++ Canvas.xcworkspace/contents.xcworkspacedata
@@ -32,7 +32,7 @@
          location = "group:NightlyTests.xctestplan">
       </FileRef>
       <FileRef
-         location = "group:StudentIPadUITests.xctestplan">
+         location = "group:IPadTests.xctestplan">
       </FileRef>
    </Group>
 </Workspace>
