+++ TestPlans/NightlyTests.xctestplan
@@ -102,6 +102,9 @@
       }
     },
     {
+      "skippedTests" : [
+        "IPadAssignmentsTest"
+      ],
       "target" : {
         "containerPath" : "container:Student.xcodeproj",
         "identifier" : "7DC5716F22F8B16200F2E9BB",
@@ -109,6 +112,10 @@
       }
     },
     {
+      "skippedTests" : [
+        "IPadSpeedGraderTests",
+        "SpeedGraderIPadTests"
+      ],
       "target" : {
         "containerPath" : "container:..\/rn\/Teacher\/ios\/Teacher.xcodeproj",
         "identifier" : "E8315D3522FC9E60003CD27C",
