--- TestPlans/StudentIPadUITests.xctestplan
@@ -1,29 +1,0 @@
-{
-  "configurations" : [
-    {
-      "id" : "B5F087BE-B1CA-4C6D-B8F0-D00717E384B2",
-      "name" : "Configuration 1",
-      "options" : {
-
-      }
-    }
-  ],
-  "defaultOptions" : {
-    "codeCoverage" : false,
-    "targetForVariableExpansion" : {
-      "containerPath" : "container:Student.xcodeproj",
-      "identifier" : "492704721CF4AF7100C631C6",
-      "name" : "Student"
-    }
-  },
-  "testTargets" : [
-    {
-      "target" : {
-        "containerPath" : "container:Student.xcodeproj",
-        "identifier" : "E8F68831239582B900A3F1CA",
-        "name" : "StudentIPadUITests"
-      }
-    }
-  ],
-  "version" : 1
-}
